# Conflicts:
#	.idea/misc.xml
#	app/src/main/res/values/strings.xml

+++ app/src/test/java/comp3350/timeSince/tests/persistence/UserPersistenceTest.java
@@ -1,41 +1,61 @@
 package comp3350.timeSince.tests.persistence;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
-import static org.junit.Assert.*;
+import org.junit.rules.ExpectedException;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Date;
 import java.util.List;
 
-import comp3350.timeSince.application.Services;
+import comp3350.timeSince.business.exceptions.DuplicateUserException;
+import comp3350.timeSince.business.exceptions.UserNotFoundException;
 import comp3350.timeSince.objects.UserDSO;
 import comp3350.timeSince.persistence.IUserPersistence;
+import comp3350.timeSince.persistence.fakes.UserPersistence;
 
 public class UserPersistenceTest {
 
     private IUserPersistence userDatabase;
     private UserDSO user1, user2, user3;
     private List<UserDSO> userList;
+    private Date defaultDate;
+
+    @Rule
+    public ExpectedException exceptionRule;
 
     @Before
     public void setUp() {
-        userDatabase = Services.getUserPersistence();
-        user1 = new UserDSO("uid1", "hash1");
-        user2 = new UserDSO("uid2", "hash2");
-        user3 = new UserDSO("uid3", "hash3");
-        userList = new ArrayList<>();
-        userList.add(user2);
-        userList.add(user3);
-        userList.add(user1);
+        userDatabase = new UserPersistence();
+        defaultDate = new Date(System.currentTimeMillis());
+        user1 = new UserDSO("uid1", defaultDate, "hash1");
+        user2 = new UserDSO("uid2", defaultDate, "hash2");
+        user3 = new UserDSO("uid3", defaultDate, "hash3");
+        userList = new ArrayList<>(Arrays.asList(user1, user2, user3));
     }
 
     @After
     public void tearDown() {
-        userDatabase.deleteUser(user1);
-        userDatabase.deleteUser(user2);
-        userDatabase.deleteUser(user3);
-        userDatabase = null;
+        List<UserDSO> tempList = userDatabase.getUserList();
+        if (tempList.contains(user1)) {
+            userDatabase.deleteUser(user1);
+        }
+        if (tempList.contains(user2)) {
+            userDatabase.deleteUser(user2);
+        }
+        if (tempList.contains(user3)) {
+            userDatabase.deleteUser(user3);
+        }
     }
 
     @Test
@@ -43,20 +63,17 @@ public class UserPersistenceTest {
         assertNotNull("Newly created database object should not be null",
                 userDatabase);
         assertEquals("Newly created database should have no users",
-                userDatabase.numUsers(), 0);
+                0, userDatabase.numUsers());
         userDatabase.insertUser(user1);
         userDatabase.insertUser(user2);
         userDatabase.insertUser(user3);
         List<UserDSO> actual = userDatabase.getUserList();
+        assertEquals("Size of database should be 3", 3, actual.size());
         assertTrue("Database should contain user1", actual.contains(user1));
         assertTrue("Database should contain user2", actual.contains(user2));
         assertTrue("Database should contain user3", actual.contains(user3));
         assertTrue("Database should have all existing users",
                 actual.containsAll(userList));
-        assertTrue("Database should have all existing users",
-                userList.containsAll(actual));
-        assertFalse("Database should not contain a user that does not exist",
-                actual.contains(new UserDSO("uid4", "password")));
     }
 
     @Test
@@ -73,7 +90,7 @@ public class UserPersistenceTest {
     @Test
     public void testInsertUser() {
         assertEquals("Size of database should be 0",
-                userDatabase.numUsers(), 0);
+                0, userDatabase.numUsers());
         assertNotNull(userDatabase.insertUser(user1));
         assertEquals("Size of database should be 1",
                 1, userDatabase.numUsers());
@@ -85,10 +102,14 @@ public class UserPersistenceTest {
         userDatabase.insertUser(user3);
         assertEquals("Size of database should be 3",
                 3, userDatabase.numUsers());
+    }
 
+    @Test(expected = DuplicateUserException.class)
+    public void testInsertUserException() {
         userDatabase.insertUser(user1);
-        assertEquals("Size of database should be 3",
-                3, userDatabase.numUsers());
+        userDatabase.insertUser(user1);
+        assertEquals("Should not be able to insert a duplicate user",
+                1, userDatabase.numUsers());
     }
 
     @Test
@@ -118,24 +139,32 @@ public class UserPersistenceTest {
         assertEquals("Size of database should be 3", 3,
                 userDatabase.numUsers());
 
-        userDatabase.deleteUser(user2);
+        assertEquals("If user exists, return the user that was deleted",
+                user1, userDatabase.deleteUser(user1));
+
         assertEquals("Size of database should be 2", 2,
                 userDatabase.numUsers());
 
         assertNull("Deleted user should no longer be in database",
-                userDatabase.getUserByID(user2.getID()));
+                userDatabase.getUserByID(user1.getID()));
+    }
 
-        assertEquals("If user exists, return the user that was deleted",
-                user1, userDatabase.deleteUser(user1));
+    @Test(expected = UserNotFoundException.class)
+    public void testDeleteUserException() {
+        userDatabase.insertUser(user1);
+        userDatabase.insertUser(user2);
+        userDatabase.insertUser(user3);
+        userDatabase.deleteUser(new UserDSO("uid4", defaultDate, "password"));
+        assertEquals("Size of database should be 3", 3, userDatabase.numUsers());
+    }
 
-        assertNull("Shouldn't be able to delete a user that doesn't exist",
-                userDatabase.deleteUser(new UserDSO("uid4", "password")));
-
-        assertEquals("Size of database should be 1", 1, userDatabase.numUsers());
-
-        userDatabase.deleteUser(user3);
-        assertEquals("Size of database should be 0", 0, userDatabase.numUsers());
-        assertNull("Should return null when database is empty",
-                userDatabase.deleteUser(user1));
+    @Test
+    public void testIsUnique() {
+        userDatabase.insertUser(user1);
+        userDatabase.insertUser(user2);
+        assertTrue("A unique user should be considered unique",
+                userDatabase.isUnique(user3.getID()));
+        assertFalse("User should not be unique if one already exists",
+                userDatabase.isUnique(user1.getID()));
     }
 }
